pyproject.toml
[project]
name = "annex"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=3.0.2",
    "python-decouple>=3.8",
]
